nothing of note

+++ src/test.cpp
@@ -28,8 +28,8 @@ int main(void){
     a +=b;
     std::cout << a.info() << valid::max<int8_t, int>(a) << N << valid::min<int8_t, int>(a) << N;
 
-    int8_t out = base_convert<int16_t>(-120, 2,3);
-    std::cout << N << (int)out << N << BITS(out)<< N << BITS((int16_t)-120);
+    int16_t out = base_convert<int16_t>(-120, 2,3);
+    std::cout << N << (int)out << N << BITS(out)<< N << BITS((int8_t)-120);
 
 
 }
